comprehensive Exploratory Data Analysis (EDA) on the Titanic dataset
The analysis includes statistical summaries, data cleaning, and a variety of visualizations using Pandas, Seaborn, and Matplotlib. Visuals such as histograms, boxplots, and bar charts were created to explore relationships between survival status and key features like age, gender, fare, and pclass. Observations and insights were added for each chart. The goal was to identify meaningful patterns and trends that contribute to survival on the Titanic.